+++ src/main/java/sg/com/petpal/petpal/model/ChatMessage.java
@@ -1,4 +1,4 @@
-package sg.com.petpal.petpal;
+package sg.com.petpal.petpal.model;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
