+++ src/main/java/sg/com/petpal/petpal/repository/ChatMessageRepository.java
@@ -1,8 +1,10 @@
-package sg.com.petpal.petpal;
+package sg.com.petpal.petpal.repository;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import sg.com.petpal.petpal.model.ChatMessage;
 
 public interface ChatMessageRepository extends JpaRepository<ChatMessage, UUID> {
